Change data types, remove duplicate files, parsing issue

+++ sqlcommon/applog.h
@@ -1,6 +1,18 @@
-
-// AppLog - Application logger class
-// Copyright (c) 2012 SQLines. All rights reserved
+/** 
+ * Copyright (c) 2016 SQLines
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 #ifndef sqlines_applog_h
 #define sqlines_applog_h
@@ -53,7 +65,7 @@ public:
     void LogConsole(const char *format, ...);
     // Write to log file only
     void LogFile(const char *format, ...);
-	void LogFile(const char *data, int len);
+	void LogFile(const char *data, int64_t len);
 
 	// Write to log file
     void LogFileVaList(const char *format, va_list args);
